ihep-application/applications/backend/integration-gateway/requirements.txt
# IHEP Integration Gateway - Python Dependencies
# Author: Jason M Jarmacz | Evolution Strategist | jason@ihep.app
# Co-Author: Claude by Anthropic

Flask>=3.0.0,<4.0.0
flask-cors>=4.0.0,<5.0.0
flask-limiter>=3.5.0,<4.0.0
requests>=2.31.0,<3.0.0
urllib3>=2.0.0,<3.0.0
PyYAML>=6.0.1,<7.0.0
google-cloud-secret-manager>=2.18.0,<3.0.0
google-cloud-pubsub>=2.19.0,<3.0.0
google-cloud-bigquery>=3.14.0,<4.0.0
hl7apy>=1.3.4,<2.0.0
fhir.resources>=7.1.0,<8.0.0
jsonschema>=4.20.0,<5.0.0
cryptography>=41.0.0,<43.0.0
gunicorn>=21.2.0,<23.0.0
python-dateutil>=2.8.2,<3.0.0
